perf: rather than copy frames alternate between alpha and beta frames
Frames do not need to be copied from one to another. At any time, we only require the previous and
current frames.Therefore, we can simply alternate between one matrix being the current frame for an
iteration. Then, on the subseuquent iteration, it becomes the previous frame, then becoming the
current frame again on the next iteration, etc. This results in an approximately 10% speedup.